fix deletion of snapshot artifact
git-svn-id: https://svn.apache.org/repos/asf/archiva/trunk@1351877 13f79535-47bb-0310-9956-ffa450edef68

+++ archiva-modules/archiva-web/archiva-rest/archiva-rest-services/src/main/java/org/apache/archiva/rest/services/DefaultRepositoriesService.java
@@ -98,6 +98,7 @@ import java.util.Collection;
 import java.util.Collections;
 import java.util.Date;
 import java.util.List;
+import java.util.Set;
 import java.util.TimeZone;
 
 /**
@@ -653,6 +654,8 @@ public class DefaultRepositoriesService
 
         // TODO more control on artifact fields
 
+        boolean snapshotVersion = VersionUtil.isSnapshot( artifact.getVersion() );
+
         RepositorySession repositorySession = repositorySessionFactory.createSession();
         try
         {
@@ -670,6 +673,13 @@ public class DefaultRepositoriesService
 
             ManagedRepositoryContent repository = repositoryFactory.getManagedRepositoryContent( repositoryId );
 
+            ArtifactReference artifactReference = new ArtifactReference();
+            artifactReference.setArtifactId( artifact.getArtifactId() );
+            artifactReference.setGroupId( artifact.getGroupId() );
+            artifactReference.setVersion( artifact.getVersion() );
+            artifactReference.setClassifier( artifact.getClassifier() );
+            artifactReference.setType( artifact.getPackaging() );
+
             MetadataRepository metadataRepository = repositorySession.getRepository();
 
             String path = repository.toMetadataPath( ref );
@@ -681,12 +691,7 @@ public class DefaultRepositoriesService
                     throw new ArchivaRestServiceException( "You must configure a type/packaging when using classifier",
                                                            400, null );
                 }
-                ArtifactReference artifactReference = new ArtifactReference();
-                artifactReference.setArtifactId( artifact.getArtifactId() );
-                artifactReference.setGroupId( artifact.getGroupId() );
-                artifactReference.setVersion( artifact.getVersion() );
-                artifactReference.setClassifier( artifact.getClassifier() );
-                artifactReference.setType( artifact.getPackaging() );
+
                 repository.deleteArtifact( artifactReference );
 
             }
@@ -705,8 +710,10 @@ public class DefaultRepositoriesService
 
                 // TODO: this should be in the storage mechanism so that it is all tied together
                 // delete from file system
-                repository.deleteVersion( ref );
-
+                if ( !snapshotVersion )
+                {
+                    repository.deleteVersion( ref );
+                }
                 File metadataFile = getMetadata( targetPath.getAbsolutePath() );
                 ArchivaRepositoryMetadata metadata = getMetadata( metadataFile );
 
@@ -715,6 +722,16 @@ public class DefaultRepositoriesService
             Collection<ArtifactMetadata> artifacts =
                 metadataRepository.getArtifacts( repositoryId, artifact.getGroupId(), artifact.getArtifactId(),
                                                  artifact.getVersion() );
+
+            if ( snapshotVersion )
+            {
+                Set<ArtifactReference> related = repository.getRelatedArtifacts( artifactReference );
+                log.debug( "related: {}", related );
+                for ( ArtifactReference artifactRef : related )
+                {
+                    repository.deleteArtifact( artifactRef );
+                }
+            }
 
             for ( ArtifactMetadata artifactMetadata : artifacts )
             {
